Adds getCampaignActivityQuery
