+++ client/src/Components/SignupList/SignupListContainer.js
@@ -5,46 +5,37 @@ import queryString from 'query-string';
 import HttpRequest from '../HttpRequest';
 import SignupListItem from './SignupListItem';
 import helpers from '../../helpers';
-import config from '../../config';
 
-/**
- * @return {Object}
- */
-function getSignupsQuery(userId) {
+const SignupListContainer = ({ userId }) => {
   const query = { include: 'posts', orderBy: 'id,desc' };
   if (userId) {
     query['filter[northstar_id]'] = userId;
   }
-  const clientQuery = queryString.parse(window.location.search);
-  if (clientQuery.skip) {
-    // Rogue API expects a page parameter for current page of results.
-    query.page = (clientQuery.skip / config.resultsPageSize) + 1;
-  }
-  if (clientQuery.source) {
-    query['filter[source]'] = clientQuery.source;
-  }
-  return query;
-}
 
-const SignupListContainer = props => (
-  <Grid>
-    <HttpRequest path={helpers.getSignupsPath()} query={getSignupsQuery(props.userId)} description="signups">
-      {res => (
-        <Table hover>
-          <tbody>
-            <Row componentClass="tr" key="header">
-              <Col md={3} componentClass="th">Created</Col>
-              <Col md={3} componentClass="th">User</Col>
-              <Col md={2} componentClass="th">Campaign</Col>
-              <Col md={6} componentClass="th" />
-            </Row>
-            {res.data.map(signup => <SignupListItem signup={signup} key={signup.id} />)}
-          </tbody>
-        </Table>
-      )}
-    </HttpRequest>
-  </Grid>
-);
+  return (
+    <Grid>
+      <HttpRequest
+        path={helpers.getSignupsPath()}
+        query={helpers.getCampaignActivityQuery(queryString.parse(window.location.search), query)}
+        description="signups"
+      >
+        {res => (
+          <Table hover>
+            <tbody>
+              <Row componentClass="tr" key="header">
+                <Col md={3} componentClass="th">Created</Col>
+                <Col md={3} componentClass="th">User</Col>
+                <Col md={2} componentClass="th">Campaign</Col>
+                <Col md={6} componentClass="th" />
+              </Row>
+              {res.data.map(signup => <SignupListItem signup={signup} key={signup.id} />)}
+            </tbody>
+          </Table>
+        )}
+      </HttpRequest>
+    </Grid>
+  );
+};
 
 SignupListContainer.propTypes = {
   userId: PropTypes.string,
