+++ client/src/helpers.js
@@ -3,6 +3,26 @@ const querystring = require('querystring');
 const config = require('./config');
 
 /**
+ * @param {Object} clientQuery
+ * @param {Object} apiQuery
+ * @return {Object}
+ */
+function getCampaignActivityQuery(clientQuery, apiQuery = {}) {
+  const query = {};
+  if (clientQuery.skip) {
+    // Rogue API expects a page parameter for current page of results.
+    query.page = (clientQuery.skip / config.resultsPageSize) + 1;
+  }
+  if (clientQuery.source) {
+    query['filter[source]'] = clientQuery.source;
+  }
+  if (clientQuery.type) {
+    query['filter[type]'] = clientQuery.type;
+  }
+  return Object.assign(apiQuery, query);
+}
+
+/**
  * @param {Array} defaultTopicTriggers
  * @return {Object}
  */
@@ -69,6 +89,7 @@ module.exports = {
   getBroadcastsPath: function getBroadcastsPath() {
     return 'broadcasts';
   },
+  getCampaignActivityQuery,
   getCampaignByIdPath: function getCampaignByIdPath(campaignId) {
     return `${this.getCampaignsPath()}/${campaignId}`;
   },
